Fix a majority of warnings in the Rust wrapper.
